# docs/swagger.yaml
definitions:
...
  controller.ErrorResponse:
    properties:
      msg:
        type: string
    type: object
...
  dto.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_order:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
      product_count:
        type: integer
      slug:
        type: string
      updated_at:
        type: string
    type: object
  dto.Product:
    properties:
      attributes:
        additionalProperties: {}
        type: object
      categories:
        items:
          $ref: '#/definitions/dto.Category'
        type: array
      cost_price:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      sale_price:
        type: number
      sku:
        type: string
      status:
        type: string
      stock_quantity:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.ProductInput:
    properties:
      attributes:
        additionalProperties: {}
        type: object
      category_ids:
        items:
          type: string
        minItems: 1
        type: array
      cost_price:
        type: number
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sale_price:
        type: number
      sku:
        type: string
      status:
        enum:
        - active
        - inactive
        - out_of_stock
        type: string
      stock_quantity:
        minimum: 0
        type: integer
    required:
    - category_ids
    - name
    - price
    - sku
    - status
    - stock_quantity
    type: object
...
info:
  contact: {}
  description: Fiber go web framework based REST API boilerplate
  title: Fiber Go API
  version: "1.0"
paths:
...
  /api/v1/products:
    get:
      consumes:
      - application/json
      description: List all products with optional filtering and pagination.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search term for name, SKU or description
        in: query
        name: search
        type: string
      - description: Filter by category ID (UUID format)
        in: query
        name: category_id
        type: string
      - description: Filter by status (active, inactive, out_of_stock)
        in: query
        name: status
        type: string
      - description: Filter by minimum price
        in: query
        name: min_price
        type: number
      - description: Filter by maximum price
        in: query
        name: max_price
        type: number
      - description: Filter by minimum stock quantity
        in: query
        name: min_stock
        type: integer
      - description: Filter by maximum stock quantity
        in: query
        name: max_stock
        type: integer
      - description: Filter by creation date (RFC3339 format)
        in: query
        name: created_after
        type: string
      - description: Filter by creation date (RFC3339 format)
        in: query
        name: created_before
        type: string
      - description: Sort field (name, price, created_at, stock_quantity)
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Product'
            type: array
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: list products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Create a new product.
      parameters:
      - description: Create new product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/model.ProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Product'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: create a new product
      tags:
      - Product
  /api/v1/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product.
      parameters:
      - description: Product ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema: {}
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: delete a product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get product details by ID.
      parameters:
      - description: Product ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Product'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: get a product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Update an existing product.
      parameters:
      - description: Product ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: Update product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/model.ProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/dto.Product'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: update a product
      tags:
      - Product
  /api/v1/products/{id}/categories:
    get:
      consumes:
      - application/json
      description: Get all categories of a product.
      parameters:
      - description: Product ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Category'
            type: array
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: get product categories
      tags:
      - Product
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"